algs/data_generation_via_llm/requirements.txt
Requests==2.31.0
mysql-connector-python